+++ ChangeLog.md
@@ -24,6 +24,9 @@ Other enhancements:
   allow, on Windows, the MSYS2 environment to be specified. The default
   environment is still `MINGW64` on 64-bit Windows and `MINGW32` on 32-bit
   Windows.
+* In YAML configuration files, the `global-hints-location` key is introduced to
+  allow the location of the global hints YAML specification file to be
+  specified.
 
 Bug fixes: